Move service creation function from api to new folder
