+++ tests/core/TestTaskService.cpp
@@ -4,7 +4,7 @@
 
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
-#include <core/api/TODOList.h>
+#include <core/service_creation/TODOList.h>
 #include "core/utils/data_transfer/TaskDTOConverter.h"
 #include "core/utils/task_io/ConsoleTaskIO.h"
 #include "mocks/CoreMocks.h"
