+++ tests/core/persistence/TestIostreamModelPersister.cpp
@@ -9,7 +9,7 @@
 #include <gmock/gmock.h>
 
 
-#include "core/api/TODOList.h"
+#include "core/service_creation/TODOList.h"
 using ::testing::NiceMock;
 using ::testing::Return;
 using ::testing::Truly;
