+++ tests/core/test_main.cpp
@@ -2,7 +2,7 @@
 // Created by denis on 14.08.20.
 //
 #include <gtest/gtest.h>
-#include "core/api/TODOList.h"
+#include "core/service_creation/TODOList.h"
 
 int main(int argc, char** argv) {
     auto srv = todo_list::createService();
